blunt copy of branches/0.60.x
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/trunk@3640 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ source/texk/kpathsea/xrealloc.c
@@ -35,8 +35,8 @@ xrealloc (void *old_ptr, size_t size)
                pointers and unsigned had different sizes.  Since the info
                is of little or no value anyway, just don't print it.  */
             fprintf(stderr,
-                    "fatal: memory exhausted (realloc of %u bytes).\n",
-                    size);
+                    "fatal: memory exhausted (realloc of %lu bytes).\n",
+                    (unsigned long)size);
             exit(EXIT_FAILURE);
         }
     }
